refactor useAttachmentLoader
- more generic attachment loader for re-use
- consumer can call dispatch
- move loadables setup to SceneManagerXR
